+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# NEXT UP
+# 3.7.10
 
 - Feat Trait advancement for Custom Lineage when importing a character. @gandamir
 - Optional Class Features would import without level restrictions since class feature refactor. (v3.7.0) @gandamir